fit pt train and model.__init__

+++ deepmd/pt/train/training.py
@@ -941,7 +941,9 @@ class Trainer:
                 continue
             if self.multi_task:
                 chosen_index_list = dp_random.choice(
-                    np.arange(self.num_model),  # pylint: disable=no-explicit-dtype
+                    np.arange(
+                        self.num_model, dtype=np.int32
+                    ),  # int32 should be enough...
                     p=np.array(self.model_prob),
                     size=self.world_size,
                     replace=True,
